Complete the JPush, you can enter the MainActivity by click the message push.
