+++ JPushLibrary/AndroidManifest.xml
@@ -32,10 +32,6 @@
 
         <!-- For test only 测试的主程序-->
         <activity android:name="com.example.jpushdemo.MainActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <!-- For test only 测试高级功能 -->
         <activity android:name="com.example.jpushdemo.PushSetActivity"/>
@@ -131,23 +127,6 @@
         <receiver
             android:name="cn.jpush.android.service.AlarmReceiver"
             android:exported="false" />
-
-        <!-- User defined.  For test only  用户自定义的广播接收器-->
-        <receiver
-            android:name="com.example.jpushdemo.MyReceiver"
-            android:enabled="true"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="cn.jpush.android.intent.REGISTRATION" /> <!--Required  用户注册SDK的intent-->
-                <action android:name="cn.jpush.android.intent.MESSAGE_RECEIVED" /> <!--Required  用户接收SDK消息的intent-->
-                <action android:name="cn.jpush.android.intent.NOTIFICATION_RECEIVED" /> <!--Required  用户接收SDK通知栏信息的intent-->
-                <action android:name="cn.jpush.android.intent.NOTIFICATION_OPENED" /> <!--Required  用户打开自定义通知栏的intent-->
-                <action android:name="cn.jpush.android.intent.ACTION_RICHPUSH_CALLBACK" /> <!--Optional 用户接受Rich Push Javascript 回调函数的intent-->
-                <action android:name="cn.jpush.android.intent.CONNECTION" /><!-- 接收网络变化 连接/断开 since 1.6.3 -->
-                <category android:name="io.xujiaji.hnbc" />
-            </intent-filter>
-        </receiver>
-
 
         <!-- Required  . Enable it you can get statistics data with channel -->
         <meta-data
